+++ Restructure Tests/RowDecoderTests.swift
@@ -1,18 +1,17 @@
 //
-//  StatementEncoderTests.swift
+//  RowDecoderTests.swift
 //  Restructure
 //
-//  Created by Stephen H. Gerstacker on 11/10/18.
+//  Created by Stephen H. Gerstacker on 11/30/18.
 //  Copyright © 2018 Stephen H. Gerstacker. All rights reserved.
 //
 
 import XCTest
-@testable import Restructure
 
-class StatementEncoderTests: XCTestCase {
-
-    var restructure: Restructure!
+class RowDecoderTests: XCTestCase {
     
+    var restructure: Restructure!
+
     override func setUp() {
         restructure = try! Restructure()
         try! restructure.execute(query: "CREATE TABLE foo (a INTEGER PRIMARY KEY AUTOINCREMENT, b TEXT, c REAL, d INT, e BLOB)")
@@ -23,8 +22,8 @@ class StatementEncoderTests: XCTestCase {
         restructure = nil
     }
 
-    func testSimpleEncodable() {
-        struct Foo: Encodable {
+    func testSimpleDecodable() throws {
+        struct Foo: Decodable {
             let a: Int64?
             let b: String
             let c: Double
@@ -32,15 +31,13 @@ class StatementEncoderTests: XCTestCase {
             let e: Data
         }
         
-        let statement = try! restructure.prepare(query: "INSERT INTO foo (b, c, d, e) VALUES (:b, :c, :d, :e)")
-        
-        let foo = Foo(a: nil, b: "1", c: 2.0, d: 3, e: Data(bytes: [0x4, 0x5, 0x6], count: 3))
-        
-        let encoder = StatementEncoder()
-        
-        XCTAssertNoThrow(try encoder.encode(foo, to: statement))
-        
-        var result = statement.step()
+        let insertStatement = try! restructure.prepare(query: "INSERT INTO foo (b, c, d, e) VALUES (:b, :c, :d, :e)")
+        insertStatement.bind(value: "1", for: "b")
+        insertStatement.bind(value: 2.0, for: "c")
+        insertStatement.bind(value: 3, for: "d")
+        insertStatement.bind(value: Data(bytes: [0x4, 0x5, 0x6], count: 3), for: "e")
+
+        var result = insertStatement.step()
         
         guard case .done = result else {
             XCTFail("Failed to insert data")
@@ -56,27 +53,25 @@ class StatementEncoderTests: XCTestCase {
             return
         }
         
-        XCTAssertEqual(row["b"], "1")
-        XCTAssertEqual(row["c"], 2.0)
-        XCTAssertEqual(row["d"], 3)
-        XCTAssertEqual(row["e"], Data(bytes: [0x4, 0x5, 0x6], count: 3))
+        let decoder = RowDecoder()
+        let foo = try decoder.decode(Foo.self, from: row)
+        
+        XCTAssertEqual(foo.b, "1")
+        XCTAssertEqual(foo.c, 2.0)
+        XCTAssertEqual(foo.d, 3)
+        XCTAssertEqual(foo.e, Data(bytes: [0x4, 0x5, 0x6], count: 3))
     }
     
-    func testArrayEncodable() {
-        struct Foo: Encodable {
+    func testArrayDecodable() throws {
+        struct Foo: Decodable {
             let a: Int64?
             let e: [Int]
         }
         
-        let statement = try! restructure.prepare(query: "INSERT INTO foo (e) VALUES (:e)")
+        let insertStatement = try! restructure.prepare(query: "INSERT INTO foo (e) VALUES (:e)")
+        insertStatement.bind(value: [1, 2, 3], for: "e")
         
-        let foo = Foo(a: nil, e: [1, 2, 3])
-        
-        let encoder = StatementEncoder()
-        
-        XCTAssertNoThrow(try encoder.encode(foo, to: statement))
-        
-        var result = statement.step()
+        var result = insertStatement.step()
         
         guard case .done = result else {
             XCTFail("Failed to insert data")
@@ -92,24 +87,22 @@ class StatementEncoderTests: XCTestCase {
             return
         }
         
-        XCTAssertEqual(row["e"], [1, 2, 3])
+        let decoder = RowDecoder()
+        let foo = try decoder.decode(Foo.self, from: row)
+        
+        XCTAssertEqual(foo.e, [1, 2, 3])
     }
     
-    func testMultiArrayEncodable() {
-        struct Foo: Encodable {
+    func testMultiDimensionArrayDecodable() throws {
+        struct Foo: Decodable {
             let a: Int64?
             let e: [[Int]]
         }
         
-        let statement = try! restructure.prepare(query: "INSERT INTO foo (e) VALUES (:e)")
+        let insertStatement = try! restructure.prepare(query: "INSERT INTO foo (e) VALUES (:e)")
+        insertStatement.bind(value: [[1, 2, 3], [4, 5, 6], [7, 8, 9]], for: "e")
         
-        let foo = Foo(a: nil, e: [[1, 2, 3]])
-        
-        let encoder = StatementEncoder()
-        
-        XCTAssertNoThrow(try encoder.encode(foo, to: statement))
-        
-        var result = statement.step()
+        var result = insertStatement.step()
         
         guard case .done = result else {
             XCTFail("Failed to insert data")
@@ -125,6 +118,10 @@ class StatementEncoderTests: XCTestCase {
             return
         }
         
-        XCTAssertEqual(row["e"], [[1, 2, 3]])
+        let decoder = RowDecoder()
+        let foo = try decoder.decode(Foo.self, from: row)
+        
+        XCTAssertEqual(foo.e, [[1, 2, 3], [4, 5, 6], [7, 8, 9]])
     }
+
 }