+++ CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Created the primary `Restructure` object for maintaining SQLite databases.
 -   Created the `Statement` object for working with prepared SQLite statements.
 -   Created the `Row` object for working with resultant rows from a `Statement`.
+-   Created the `RowDecoder` and `StatementEncoder` for using Swift's `Decodable` protocols.